refactor: Players 클래스 생성, Users 메소드 일부를 대체
- Users는 Dealer와 Player를 구분 없이 담고 있어 필요시 필터링 하는 과정이 있었습니다. 이에 Player만을 담는 Players를 생성하여 이 과정을 생략할 수 있도록 대체하였습니다.
