+++ src/main/java/blackjack/controller/BlackJackController.java
@@ -1,40 +1,43 @@
 package blackjack.controller;
 
+import blackjack.domain.card.CardDeck;
 import blackjack.domain.user.Dealer;
 import blackjack.domain.user.Player;
+import blackjack.domain.user.Players;
 import blackjack.domain.user.Users;
-import blackjack.domain.card.CardDeck;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
 
 public class BlackJackController {
     private static final int INITIAL_DRAW_CARD_NUMBER = 2;
 
-    private final Users users;
+    private final Players players;
     private final Dealer dealer;
     private final CardDeck cardDeck;
+    private final Users users;
 
     public BlackJackController() {
         this.dealer = new Dealer();
         this.cardDeck = CardDeck.createDeck();
-        this.users = new Users(dealer, InputView.scanPlayerNames());
+        this.players = Players.of(InputView.scanPlayerNames());
+        this.users = new Users(this.dealer, this.players);
     }
 
     public void run() {
         initialDealingCards(users);
-        OutputView.printInitialComment(users);
+        OutputView.printInitialComment(dealer, players);
         OutputView.printCardsOfUsersWithScore(users);
 
         if (dealer.isBlackJack()) {
-            OutputView.printResult(users.generateResultsMapAgainstDealer());
+            OutputView.printResult(players.generateResultsMapAgainstDealer(dealer));
             return;
         }
 
-        users.getPlayers().forEach(this::playGameForEachPlayer);
+        players.players().forEach(this::playGameForEachPlayer);
         drawCardsOfDealerUntilOver16Score();
 
         OutputView.printCardsOfUsersWithScore(users);
-        OutputView.printResult(users.generateResultsMapAgainstDealer());
+        OutputView.printResult(players.generateResultsMapAgainstDealer(dealer));
     }
 
     private void drawCardsOfDealerUntilOver16Score() {
@@ -46,8 +49,7 @@ public class BlackJackController {
 
     private void initialDealingCards(Users users) {
         for (int i = 0; i < INITIAL_DRAW_CARD_NUMBER; i++) {
-            users.gerUsers()
-                    .forEach(user -> user.addCard(cardDeck.drawCard()));
+            users.users().forEach(user -> user.addCard(cardDeck.drawCard()));
         }
     }
